Fixed some bugs on CUDA algorithm; Added more types of product for gpu; Added file to fill csv with name of the matrices; Completed script to check the if the result of the product is correct;

+++ lib/product_cpu.c
@@ -9,7 +9,7 @@ double serial_product_csr(csr_matrix* matrix, double* array, double* result) {
 
 	double start = omp_get_wtime();
 
-	int sum;
+	double sum;
 	for (int i = 0; i < matrix->M; i++) {
 		sum = 0;
 		for (int j = matrix->irp[i]; j < matrix->irp[i+1]; j++)
@@ -27,7 +27,7 @@ double serial_product_ellpack(ellpack_matrix* matrix, double* array, double* res
 
 	double start = omp_get_wtime();
 
-	int sum;
+	double sum;
 	for (int i = 0; i < matrix->M; i++) {
 		sum = 0;
 		for (int j = 0; j < matrix->maxnz; j++)
@@ -46,7 +46,7 @@ double omp_product_csr(csr_matrix* matrix, double* array, double* result) {
 	int i, j;
 	#pragma omp parallel for schedule(static) default(none) private(i, j) shared(matrix, array, result)
 	for (i = 0; i < matrix->M; i++) {
-        int sum = 0;
+        double sum = 0;
 		for (j = matrix->irp[i]; j < matrix->irp[i+1]; j++)
 			sum += matrix->as[j]*array[matrix->ja[j]];
 
@@ -64,7 +64,7 @@ double omp_product_ellpack(ellpack_matrix* matrix, double* array, double* result
 	int i, j;
 	#pragma omp parallel for schedule(static) default(none) private(i, j) shared(matrix, array, result)
 	for (i = 0; i < matrix->M; i++) {
-		int sum = 0;
+		double sum = 0;
 		for (j = 0; j < matrix->maxnz; j++) {
 			sum += matrix->as[i][j]*array[matrix->ja[i][j]];
 		}
